BE/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
requests
python-dotenv
paho-mqtt
websockets
